+++ packages/splunk_otel_flutter/splunk_otel_flutter/lib/splunk_otel_flutter.dart
@@ -92,7 +92,7 @@ export 'package:splunk_otel_flutter_platform_interface/src/model/mutable_attribu
 export 'package:splunk_otel_flutter/src/splunk_otel_flutter.dart';
 export 'package:splunk_otel_flutter/src/global_attributes.dart';
 export 'package:splunk_otel_flutter/src/custom_tracking.dart'
-    show WorkflowHandle;
+    show WorkflowHandle, ErrorSource;
 export 'package:splunk_otel_flutter/src/navigation/splunk_navigator_observer.dart'
     show
         SplunkNavigatorObserver,